+++ Nickel/Models/Content/DeckConfiguration.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace Nickel;
 
 /// <summary>
@@ -19,4 +21,20 @@ public readonly struct DeckConfiguration
 	
 	/// <summary>A localization provider for the name of the <see cref="Deck"/>.</summary>
 	public SingleLocalizationProvider? Name { get; init; }
+	
+	/// <summary>A delegate that can override the default rarity shine of a card.</summary>
+	public Func<ShineColorOverrideArgs, Color>? ShineColorOverride { get; init; }
+
+	/// <seealso cref="ShineColorOverride"/>
+	public struct ShineColorOverrideArgs
+	{
+		/// <summary>The current state of the game.</summary>
+		public required State State { get; init; }
+		
+		/// <summary>The card being rendered.</summary>
+		public required Card Card { get; init; }
+		
+		/// <summary>The default color of the rarity shine for this card.</summary>
+		public required Color DefaultShineColor { get; init; }
+	}
 }